requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
google-cloud-pubsub==2.13.0
google-cloud-bigquery==2.34.3
redis==4.3.4
apache-beam[gcp]==2.40.0
pydantic==1.9.0
python-json-logger==2.0.4
jsonschema==4.17.3
pyyaml==6.0
pytest==7.3.1
pytest-asyncio==0.20.3
pytest-cov==4.0.0
black==22.3.0
flake8==4.0.1
mypy==0.991
psutil==5.9.0
python-dotenv==0.19.2
